// aiot-frontend/src/api/generated/device-service/models/device-v1-simulate-push-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * SimulatePush - Online Debug Feature
 */
export interface DeviceV1SimulatePushRequest {
    'deviceKey'?: string;
    'operationType'?: string;
    'operationName'?: string;
    'payload'?: string;
}